Relaciones de Cliente. No sé si las relaciones deben ser bidireccionales, así que esa parte está comentada.

+++ src/main/java/Entities/HistorialPedidos.java
@@ -16,10 +16,11 @@ public class HistorialPedidos {
     @Id
     @Column(name = "idpedido")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idPedido;
+    private Long idPedido;
 
     @Column(name = "idCliente")
-    private int idCliente;
+    private Long idCliente;
+    
     @Column(name = "fechaPedido")
     private String fechaPedido;
 
@@ -32,4 +33,8 @@ public class HistorialPedidos {
     @Column(name = "subtotal")
     private String subtotal;
 
+    //No sé si la relación debe ser bidireccional o no, así que comento -Nick
+    //Relación Bidireccional OneToOne con Cliente
+    //@OneToOne(mappedBy = "historialPedidos")
+    //private Cliente cliente;
 }
